fix: user data ignition is ignored
I missed the note in the docs that the ignition config specified on
`install-flatcar` will disable the `fetch` stage in ignition. This stage
is required to read the user data ignition config from the metadata
service.

By moving our OEM ignition config to `oemfs/base/base.ign` it will
always be merged with the user data config.